Feat: Remodeling database
Deleted all the existing event tables and created a new tables like this:

Tables for crawling: FestaCrawling, MeetupCrawling, EventusCrawling, FacebookCrawling

Tables for in-house handling event: ManualEvent, UserrequestEvent, WaitingEvent

Tables for API: DevEvent

Later we will create a model to classificate which events are development related or not. For training data, non-development events are stored in the 'NotDevEvent' Table.